better axes for daily aggregation
previously plots would only display days appearing in the data set. If there were no obs for a day, this day would just be skipped. For example, if 3-12 to 3-17 had no obs, then 3-11 and 3-18 would be plotted next to each other. This lead to inconsistent month sizes in months without  a lot of observations. Now, a forced observation of 0 appears if there are no obs on that month-day.